+++ src/main/java/com/sparta/tentenbackend/domain/menu/dto/MenuOptionDto.java
@@ -1,8 +1,6 @@
 package com.sparta.tentenbackend.domain.menu.dto;
 
 import com.sparta.tentenbackend.domain.menu.entity.MenuOption;
-
-import java.math.BigInteger;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -18,17 +16,17 @@ import lombok.Setter;
 public class MenuOptionDto {
 
 
-  private UUID id;
+    private UUID id;
 
-  private String category;
-  private BigInteger price;
+    private String category;
+    private Long price;
 
-  // 생성 메서드 (엔티티 -> DTO)
-  public static MenuOptionDto fromEntity(MenuOption menuOption) {
-    return MenuOptionDto.builder()
-        .id(menuOption.getId())
-        .category(menuOption.getCategory())
-        .price(menuOption.getPrice())
-        .build();
-  }
+    // 생성 메서드 (엔티티 -> DTO)
+    public static MenuOptionDto fromEntity(MenuOption menuOption) {
+        return MenuOptionDto.builder()
+            .id(menuOption.getId())
+            .category(menuOption.getCategory())
+            .price(menuOption.getPrice())
+            .build();
+    }
 }
