+++ src/main/java/com/sparta/tentenbackend/domain/order/service/OrderRepositoryService.java
@@ -3,7 +3,6 @@ package com.sparta.tentenbackend.domain.order.service;
 import com.sparta.tentenbackend.domain.order.dto.OrderSearchRequest;
 import com.sparta.tentenbackend.domain.order.dto.TemporaryOrderRequest;
 import com.sparta.tentenbackend.domain.order.entity.Order;
-import com.sparta.tentenbackend.domain.user.entity.User;
 import java.util.UUID;
 import org.springframework.data.domain.Page;
 
@@ -16,8 +15,4 @@ public interface OrderRepositoryService {
     Page<Order> getOrderListByStoreId(UUID storeId, OrderSearchRequest orderSearchRequest);
 
     Order createTemporaryOrder(TemporaryOrderRequest req);
-
-    void acceptOrder(UUID orderId, User owner);
-
-    void rejectOrder(UUID orderId, User owner);
 }
